Fix build errors on 64bit platforms.
Thanks to jeremyw...@gmail.com:
https://code.google.com/p/anic/issues/detail?id=9

+++ src/types.cpp
@@ -1511,7 +1511,7 @@ TypeStatus::TypeStatus(Type *type, Type *retType) : type(type), retType(retType)
 TypeStatus::TypeStatus(Type *type, const TypeStatus &otherStatus) : type(type), retType(otherStatus.retType), code(NULL) {}
 TypeStatus::~TypeStatus() {}
 TypeStatus::operator Type *() const {return type;}
-TypeStatus::operator uintptr_t() const {return (unsigned int)type;}
+TypeStatus::operator uintptr_t() const {return (uintptr_t)type;}
 DataTree *TypeStatus::castCode(const Type &destType) const {
 	StdType *thisType = (StdType *)type;
 	StdType *otherType = (StdType *)(&destType);
